# packages/http-api/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Feedless HTTP API
  description: REST API for Feedless entities (auth, repository, source, document, harvest, plan, group).
  version: 1.0.0
servers:
  - url: '{scheme}://{host}/api/v1'
    description: Feedless HTTP API v1. Feedless is self-hosted — point `host` at your instance.
    variables:
      scheme:
        default: https
        enum: [https, http]
      host:
        default: feedless.org
        description: Host (and optional port) of the Feedless instance
security:
  - bearerAuth: []
paths:
...
  /repositories:
    get:
      tags: [Repositories]
      operationId: listRepositories
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: product
          in: query
          schema:
            $ref: '#/components/schemas/VerticalFilter'
        - name: visibility
          in: query
          schema:
            $ref: '#/components/schemas/Visibility'
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Repositories]
      operationId: createRepository
      summary: Create a single repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepositoryCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /repositories/count:
    get:
      tags: [Repositories]
      operationId: countRepositories
      parameters:
        - name: product
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/VerticalFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountResponse'
        default:
          $ref: '#/components/responses/Error'
  /repositories/{repositoryId}:
    parameters:
      - $ref: '#/components/parameters/RepositoryId'
    get:
      tags: [Repositories]
      operationId: getRepository
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Repositories]
      operationId: updateRepository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepositoryUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Repositories]
      operationId: deleteRepository
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /repositories/{repositoryId}/sources:
    parameters:
      - $ref: '#/components/parameters/RepositoryId'
    get:
      tags: [Sources]
      operationId: listSources
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: disabled
          in: query
          schema:
            type: boolean
        - name: like
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Sources]
      operationId: createSource
      summary: Create a single source in a repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /repositories/{repositoryId}/sources/{sourceId}:
    parameters:
      - $ref: '#/components/parameters/RepositoryId'
      - $ref: '#/components/parameters/SourceId'
    get:
      tags: [Sources]
      operationId: getSource
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Sources]
      operationId: updateSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Sources]
      operationId: deleteSource
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <UserSecret.value>` (GitHub PAT style).
        Legacy `Authentication: Bearer <…>` accepted in v1 but **deprecated**.
        TOKEN cookie still accepted for browser clients. Anonymous JWTs are rejected.
  parameters:
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 100
        default: 20
    SourceId:
      name: sourceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: >
        Any error. Every error response — 400 `BAD_REQUEST` / `VALIDATION_ERROR`,
        401 `UNAUTHORIZED`, 403 `FORBIDDEN`, 404 `NOT_FOUND`, 429 `TOO_MANY_REQUESTS`,
        500 `INTERNAL_ERROR` — carries an `ApiError` body. Since every route requires
        authentication, every operation can return 401.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    TooManyRequests:
      description: Rate limit exceeded — retry after the interval in `Retry-After`
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable code, e.g. NOT_FOUND, VALIDATION_ERROR
        message:
          type: string
          description: Human-readable summary. Not stable — do not match on it.
        corrId:
          type: string
          description: Correlation id for support/debug
        path:
          type: string
          description: Request path the error occurred on
        errors:
          type: array
          description: Per-field details, present on VALIDATION_ERROR. All failing fields, not just the first.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    CountResponse:
      type: object
      required: [count]
      properties:
        count:
          type: integer
    RepositoryListResponse:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Repository'
        hasMore:
          type: boolean
    SourceListResponse:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        hasMore:
          type: boolean
...
    Vertical:
      type: string
      description: A concrete product. Never `all` — that is a filter value only, see VerticalFilter.
      enum:
        - rssProxy
        - visualDiff
        - reader
        - feedless
        - upcoming
        - untoldNotes
        - pageChangeTracker
        - feedDump
        - digest
    VerticalFilter:
      type: string
      description: Product filter for queries. `all` matches every product.
      enum:
        - all
        - rssProxy
        - visualDiff
        - reader
        - feedless
        - upcoming
        - untoldNotes
        - pageChangeTracker
        - feedDump
        - digest
    Visibility:
      type: string
      enum:
        - private
        - public
    Retention:
      type: object
      properties:
        maxCapacity:
          type: integer
        maxAgeDays:
          type: integer
    Repository:
      type: object
      description: >
        Only intrinsic, always-present fields are required. Computed counters and
        caller-relative fields are optional and are omitted when not resolved.
      required:
        - id
        - title
        - description
        - ownerId
        - product
        - visibility
        - refreshCron
        - tags
        - createdAt
        - lastUpdatedAt
        - archived
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        shareKey:
          type: string
          description: >
            Capability secret granting access to a private repository. Returned only to
            the owner; omitted for everyone else. Treat as a credential.
        ownerId:
          type: string
          format: uuid
        product:
          $ref: '#/components/schemas/Vertical'
        visibility:
          $ref: '#/components/schemas/Visibility'
        refreshCron:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        nextUpdateAt:
          type: string
          format: date-time
        documentCount:
          type: integer
          format: int64
        archived:
          type: boolean
        pullsPerMonth:
          type: integer
        currentUserIsOwner:
          type: boolean
          description: Relative to the calling user, not a property of the repository itself
        pushNotificationsEnabled:
          type: boolean
        sourcesCount:
          type: integer
          description: Omitted when not computed for this response
        sourcesCountWithProblems:
          type: integer
          description: Omitted when not computed for this response
        disabledFrom:
          type: string
          format: date-time
    RepositoryCreate:
      type: object
      required: [product, sources, title, description]
      properties:
        product:
          $ref: '#/components/schemas/Vertical'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceCreate'
        title:
          type: string
        description:
          type: string
        retention:
          $ref: '#/components/schemas/Retention'
        refreshCron:
          type: string
        withShareKey:
          type: boolean
        pushNotificationsMuted:
          type: boolean
        visibility:
          $ref: '#/components/schemas/Visibility'
    RepositoryUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        refreshCron:
          type: string
        pushNotificationsMuted:
          type: boolean
        visibility:
          $ref: '#/components/schemas/Visibility'
        retention:
          $ref: '#/components/schemas/Retention'
    Source:
      type: object
      required: [id, title, flow]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        flow:
          $ref: '#/components/schemas/ScrapeFlow'
        disabled:
          type: boolean
        lastRefreshedAt:
          type: string
          format: date-time
        lastRecordsRetrieved:
          type: integer
        latLng:
          $ref: '#/components/schemas/GeoPoint'
        tags:
          type: array
          items:
            type: string
        recordCount:
          type: integer
          description: Omitted when not computed for this response
        lastErrorMessage:
          type: string
    SourceCreate:
      type: object
      required: [title, flow]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        flow:
          $ref: '#/components/schemas/ScrapeFlow'
        draft:
          type: boolean
        latLng:
          $ref: '#/components/schemas/GeoPoint'
        tags:
          type: array
          items:
            type: string
        lastErrorMessage:
          type: string
    SourceUpdate:
      type: object
      properties:
        title:
          type: string
        tags:
          type: array
          items:
            type: string
        disabled:
          type: boolean
        latLng:
          $ref: '#/components/schemas/GeoPoint'
        flow:
          $ref: '#/components/schemas/ScrapeFlow'
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
    ScrapeFlow:
      type: object
      required: [sequence]
      properties:
        sequence:
          type: array
          items:
            $ref: '#/components/schemas/ScrapeAction'
    ScrapeAction:
      type: object
      description: >
        Exactly one of these properties must be set — an action is a fetch, or a click, or
        an extract, never a combination. Enforced at runtime: an action with zero or more
        than one property set is rejected with 400 VALIDATION_ERROR.
      minProperties: 1
      maxProperties: 1
      properties:
        fetch:
          $ref: '#/components/schemas/HttpFetch'
        extract:
          $ref: '#/components/schemas/ScrapeExtract'
        execute:
          $ref: '#/components/schemas/PluginExecution'
        click:
          $ref: '#/components/schemas/DomElement'
        type:
          $ref: '#/components/schemas/DomActionType'
        waitFor:
          $ref: '#/components/schemas/WaitForAction'
        select:
          $ref: '#/components/schemas/DomActionSelect'
        header:
          $ref: '#/components/schemas/RequestHeader'
        purge:
          $ref: '#/components/schemas/DomElementByXPath'
    HttpFetch:
      type: object
      properties:
        get:
          $ref: '#/components/schemas/HttpGetRequest'
    HttpGetRequest:
      type: object
      required: [url]
      properties:
        url:
          $ref: '#/components/schemas/StringLiteralOrVariable'
        timeout:
          type: integer
        viewport:
          $ref: '#/components/schemas/ViewPort'
        forcePrerender:
          type: boolean
        language:
          type: string
        waitUntil:
          $ref: '#/components/schemas/PuppeteerWaitUntil'
        additionalWaitSec:
          type: integer
    ViewPort:
      type: object
      required: [width, height, isMobile, isLandscape]
      properties:
        width:
          type: integer
        height:
          type: integer
        isMobile:
          type: boolean
        isLandscape:
          type: boolean
    PuppeteerWaitUntil:
      type: string
      enum:
        - load
        - domcontentloaded
        - networkidle0
        - networkidle2
    StringLiteralOrVariable:
      type: object
      properties:
        literal:
          type: string
        variable:
          type: string
    ScrapeExtract:
      type: object
      required: [fragmentName]
      properties:
        fragmentName:
          type: string
        imageBased:
          $ref: '#/components/schemas/ScrapeBoundingBox'
        selectorBased:
          $ref: '#/components/schemas/DomExtract'
    ScrapeBoundingBox:
      type: object
      required: [boundingBox]
      properties:
        boundingBox:
          $ref: '#/components/schemas/BoundingBox'
    BoundingBox:
      type: object
      required: [x, y, w, h]
      properties:
        x:
          type: integer
        y:
          type: integer
        w:
          type: integer
        h:
          type: integer
    DomExtract:
      type: object
      required: [fragmentName, xpath, emit, uniqueBy]
      properties:
        fragmentName:
          type: string
        xpath:
          $ref: '#/components/schemas/DomElementByXPath'
        max:
          type: integer
        emit:
          type: array
          items:
            $ref: '#/components/schemas/ScrapeEmit'
        uniqueBy:
          $ref: '#/components/schemas/ScrapeEmit'
        extract:
          type: array
          items:
            $ref: '#/components/schemas/DomExtract'
    DomElementByXPath:
      type: object
      required: [value]
      properties:
        value:
          type: string
    ScrapeEmit:
      type: string
      enum:
        - text
        - html
        - pixel
        - date
    DomElement:
      type: object
      properties:
        element:
          $ref: '#/components/schemas/DomElementByNameOrXPath'
        position:
          $ref: '#/components/schemas/XyPosition'
    DomElementByNameOrXPath:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/DomElementByName'
        xpath:
          $ref: '#/components/schemas/DomElementByXPath'
    DomElementByName:
      type: object
      required: [value]
      properties:
        value:
          type: string
    XyPosition:
      type: object
      required: [x, y]
      properties:
        x:
          type: integer
        y:
          type: integer
    WaitForAction:
      type: object
      required: [element]
      properties:
        element:
          $ref: '#/components/schemas/DomElementByNameOrXPath'
    DomActionType:
      type: object
      required: [element, typeValue]
      properties:
        element:
          $ref: '#/components/schemas/DomElementByXPath'
        typeValue:
          type: string
    DomActionSelect:
      type: object
      required: [element, selectValue]
      properties:
        element:
          $ref: '#/components/schemas/DomElementByXPath'
        selectValue:
          type: string
    RequestHeader:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
    PluginExecution:
      type: object
      required: [pluginId, params]
      properties:
        pluginId:
          type: string
        params:
          $ref: '#/components/schemas/PluginExecutionParams'
    PluginExecutionParams:
      type: object
      properties:
        jsonData:
          type: string
        org_feedless_fulltext:
          $ref: '#/components/schemas/FulltextPluginParams'
        org_feedless_diff_records:
          $ref: '#/components/schemas/DiffRecordsParams'
        org_feedless_feed:
          $ref: '#/components/schemas/FeedParams'
    FulltextPluginParams:
      type: object
      required: [readability, summary, inheritParams]
      properties:
        readability:
          type: boolean
        summary:
          type: boolean
        inheritParams:
          type: boolean
        onErrorRemove:
          type: boolean
    DiffRecordsParams:
      type: object
      required: [nextItemMinIncrement, compareBy]
      properties:
        nextItemMinIncrement:
          type: number
          format: float
        compareBy:
          $ref: '#/components/schemas/CompareBy'
        inlineDiffImage:
          type: boolean
        inlineLatestImage:
          type: boolean
        inlinePreviousImage:
          type: boolean
    CompareBy:
      type: object
      required: [field]
      properties:
        fragmentNameRef:
          type: string
        field:
          type: string
    FeedParams:
      type: object
      properties:
        generic:
          $ref: '#/components/schemas/Selectors'
    Selectors:
      type: object
      required: [contextXPath, dateIsStartOfEvent, extendContext, paginationXPath, linkXPath]
      properties:
        contextXPath:
          type: string
        dateIsStartOfEvent:
          type: boolean
        dateXPath:
          type: string
        extendContext:
          type: string
        paginationXPath:
          type: string
        linkXPath:
          type: string
...